+++ src/lib.rs
@@ -1,5 +1,5 @@
 /*!
-Crate for determining **binary-based** file formats.
+Crate for determining **file formats**.
 
 ## Examples
 
@@ -36,13 +36,13 @@ Add this to your `Cargo.toml`:
 
 ```toml
 [dependencies]
-file-format = "0.10"
+file-format = "0.11"
 ```
 
 ## Features
 
-* **cfb** - Enables **Compound File Binary** formats support.
-* **zip** - Enables **ZIP** formats support.
+- **cfb** - Enables **Compound File Binary** formats support.
+- **zip** - Enables **ZIP** formats support.
 
 All these features are disabled by default.
 */
@@ -201,6 +201,7 @@ formats! {
     media_type = "application/octet-stream"
     extension = "bin"
     kind = Application
+    comment = "Default format"
 
     format = ArchivedByRobertJung
     name = "Archived by Robert Jung"
@@ -291,6 +292,12 @@ formats! {
     media_type = "application/vnd.circuitdiagram.document.main+xml"
     extension = "cddx"
     kind = Application
+
+    format = ClojureScript
+    name = "Clojure Script"
+    media_type = "text/x-clojure"
+    extension = "clj"
+    kind = Text
 
     format = CommonObjectFileFormat
     name = "Common Object File Format"
@@ -406,6 +413,18 @@ formats! {
     extension = "xar"
     kind = Application
 
+    format = ExtensibleMarkupLanguage
+    name = "Extensible Markup Language"
+    media_type = "text/xml"
+    extension = "xml"
+    kind = Text
+
+    format = ExtensibleStylesheetLanguageTransformations
+    name = "Extensible Stylesheet Language Transformations"
+    media_type = "application/xslt+xml"
+    extension = "xsl"
+    kind = Application
+
     format = FastTracker2ExtendedModule
     name = "FastTracker 2 Extended Module"
     media_type = "audio/x-xm"
@@ -458,6 +477,12 @@ formats! {
     name = "Game Boy ROM"
     media_type = "application/x-gameboy-rom"
     extension = "gb"
+    kind = Application
+
+    format = GeographyMarkupLanguage
+    name = "Geography Markup Language"
+    media_type = "application/gml+xml"
+    extension = "gml"
     kind = Application
 
     format = GlTransmissionFormatBinary
@@ -513,6 +538,12 @@ formats! {
     media_type = "image/heif-sequence"
     extension = "heifs"
     kind = Image
+
+    format = HypertextMarkupLanguage
+    name = "HyperText Markup Language"
+    media_type = "text/html"
+    extension = "html"
+    kind = Text
 
     format = IccProfile
     name = "ICC Profile"
@@ -598,6 +629,12 @@ formats! {
     extension = "jxl"
     kind = Image
 
+    format = KeyholeMarkupLanguage
+    name = "Keyhole Markup Language"
+    media_type = "application/vnd.google-earth.kml+xml"
+    extension = "kml"
+    kind = Application
+
     format = KeyholeMarkupLanguageZipped
     name = "Keyhole Markup Language Zipped"
     media_type = "application/vnd.google-earth.kmz"
@@ -645,6 +682,12 @@ formats! {
     media_type = "application/x-lua-bytecode"
     extension = "luac"
     kind = Application
+
+    format = LuaScript
+    name = "Lua Script"
+    media_type = "text/x-lua"
+    extension = "lua"
+    kind = Text
 
     format = Lz4
     name = "LZ4"
@@ -844,6 +887,18 @@ formats! {
     extension = "mid"
     kind = Audio
 
+    format = Musicxml
+    name = "MusicXML"
+    media_type = "application/vnd.recordare.musicxml+xml"
+    extension = "musicxml"
+    kind = Application
+
+    format = MusicxmlZipped
+    name = "MusicXML Zipped"
+    media_type = "application/vnd.recordare.musicxml"
+    extension = "mxl"
+    kind = Application
+
     format = NikonElectronicFile
     name = "Nikon Electronic File"
     media_type = "image/x-nikon-nef"
@@ -1024,6 +1079,12 @@ formats! {
     extension = "key"
     kind = Application
 
+    format = PerlScript
+    name = "Perl Script"
+    media_type = "text/x-perl"
+    extension = "pl"
+    kind = Text
+
     format = PgpMessage
     name = "PGP Message"
     media_type = "application/pgp"
@@ -1054,6 +1115,13 @@ formats! {
     extension = "asc"
     kind = Application
 
+    format = PlainText
+    name = "Plain Text"
+    media_type = "text/plain"
+    extension = "txt"
+    kind = Text
+    comment = "UTF-8–encoded"
+
     format = PortableDocumentFormat
     name = "Portable Document Format"
     media_type = "application/pdf"
@@ -1078,6 +1146,12 @@ formats! {
     extension = "ps"
     kind = Application
 
+    format = PythonScript
+    name = "Python Script"
+    media_type = "text/x-script.python"
+    extension = "py"
+    kind = Text
+
     format = QualcommPureVoice
     name = "Qualcomm PureVoice"
     media_type = "audio/qcelp"
@@ -1089,6 +1163,12 @@ formats! {
     media_type = "image/vnd.radiance"
     extension = "hdr"
     kind = Image
+
+    format = ReallySimpleSyndication
+    name = "Really Simple Syndication"
+    media_type = "application/rss+xml"
+    extension = "rss"
+    kind = Application
 
     format = RedHatPackageManager
     name = "Red Hat Package Manager"
@@ -1107,6 +1187,18 @@ formats! {
     media_type = "application/vnd.rar"
     extension = "rar"
     kind = Application
+
+    format = RubyScript
+    name = "Ruby Script"
+    media_type = "text/x-ruby"
+    extension = "rb"
+    kind = Text
+
+    format = ScalableVectorGraphics
+    name = "Scalable Vector Graphics"
+    media_type = "image/svg+xml"
+    extension = "svg"
+    kind = Image
 
     format = Screamtracker3Module
     name = "ScreamTracker 3 Module"
@@ -1130,6 +1222,18 @@ formats! {
     name = "Shapefile"
     media_type = "application/x-esri-shape"
     extension = "shp"
+    kind = Application
+
+    format = ShellScript
+    name = "Shell Script"
+    media_type = "text/x-shellscript"
+    extension = "sh"
+    kind = Text
+
+    format = SimpleObjectAccessProtocol
+    name = "Simple Object Access Protocol"
+    media_type = "application/soap+xml"
+    extension = "soap"
     kind = Application
 
     format = Sketchup
@@ -1204,6 +1308,12 @@ formats! {
     extension = "3mf"
     kind = Application
 
+    format = ToolCommandLanguageScript
+    name = "Tool Command Language Script"
+    media_type = "text/x-tcl"
+    extension = "tcl"
+    kind = Text
+
     format = Truetype
     name = "TrueType"
     media_type = "font/ttf"
@@ -1221,6 +1331,18 @@ formats! {
     media_type = "application/x-compress"
     extension = "Z"
     kind = Application
+
+    format = Vcalendar
+    name = "vCalendar"
+    media_type = "text/calendar"
+    extension = "ics"
+    kind = Text
+
+    format = Vcard
+    name = "vCard"
+    media_type = "text/vcard"
+    extension = "vcf"
+    kind = Text
 
     format = VirtualboxVirtualDiskImage
     name = "VirtualBox Virtual Disk Image"
@@ -1366,7 +1488,7 @@ impl FileFormat {
     ///
     /// # Examples
     ///
-    /// Detects from the first bytes of a PNG file:
+    /// Detects from the first bytes of a **PNG** file:
     ///
     /// ```rust
     /// use file_format::FileFormat;
@@ -1397,8 +1519,8 @@ impl FileFormat {
     /// ```rust
     /// use file_format::FileFormat;
     ///
-    /// let format = FileFormat::from_file("fixtures/video/sample.mkv")?;
-    /// assert_eq!(format, FileFormat::MatroskaVideo);
+    /// let format = FileFormat::from_file("fixtures/text/sample.txt")?;
+    /// assert_eq!(format, FileFormat::PlainText);
     /// # Ok::<(), std::io::Error>(())
     ///```
     #[inline]
@@ -1419,18 +1541,23 @@ impl FileFormat {
     ///```
     pub fn from_reader<R: Read + Seek>(reader: R) -> Result<Self> {
         let mut reader = BufReader::with_capacity(36870, reader);
-        Ok(Self::from_signature(reader.fill_buf()?)
-            .map(|format| match format {
+        Ok(if reader.fill_buf()?.is_empty() {
+            Self::default()
+        } else if let Some(format) = Self::from_signature(reader.buffer()) {
+            match format {
                 #[cfg(feature = "cfb")]
-                Self::CompoundFileBinary => Self::from_cfb(reader).unwrap_or_default(),
-                Self::MatroskaVideo => Self::from_mkv(reader).unwrap_or_default(),
-                Self::MsDosExecutable => Self::from_ms_dos_executable(reader).unwrap_or_default(),
-                Self::PortableDocumentFormat => Self::from_pdf(reader).unwrap_or_default(),
+                Self::CompoundFileBinary => Self::from_cfb(&mut reader).unwrap_or_default(),
+                Self::ExtensibleMarkupLanguage => Self::from_xml(&mut reader).unwrap_or_default(),
+                Self::MatroskaVideo => Self::from_mkv(&mut reader).unwrap_or_default(),
+                Self::MsDosExecutable => Self::from_ms_dos_exe(&mut reader).unwrap_or_default(),
+                Self::PortableDocumentFormat => Self::from_pdf(&mut reader).unwrap_or_default(),
                 #[cfg(feature = "zip")]
-                Self::Zip => Self::from_zip(reader).unwrap_or_default(),
+                Self::Zip => Self::from_zip(&mut reader).unwrap_or_default(),
                 _ => format,
-            })
-            .unwrap_or_default())
+            }
+        } else {
+            Self::from_plain_text(&mut reader).unwrap_or_default()
+        })
     }
 }
 
@@ -1456,7 +1583,7 @@ impl From<&[u8]> for FileFormat {
     }
 }
 
-/// A kind of [`FileFormat`] according to the media type.
+/// A kind of [`FileFormat`] according to the **media type**.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum Kind {
     /// Data to be processed by some type of application program.
@@ -1469,6 +1596,8 @@ pub enum Kind {
     Image,
     /// 3D model.
     Model,
+    /// Text.
+    Text,
     /// Video.
     Video,
 }